done 85% of s1
